requirements.txt
flask==1.1.2
flask-wtf==0.14.3
Flask-Session==0.3.2
flask-bootstrap==3.3.7.1
requests==2.24.0
plotly==4.9.0
matplotlib==3.3.1
Pillow==7.2.0